TOSgui - TMO preview and bug fixes

+++ TMOgui/TMOGUIToneMappingChooser.cpp
@@ -4,7 +4,7 @@
 #include "../tmolib/TMO.h"
 #include "TMOGUIToneMappingChooser.h"
 #include "TMOGUIToneMapping.h"
-#include "TMOGUIImage.h"
+#include "TMOGUIPreviewImage.h"
 #include "TMOGUITransformation.h"
 #include "TMOGUIBitmap.h"
 #include <qpushbutton.h>
@@ -14,20 +14,24 @@
 #include <qpixmap.h>
 #include <qtemporarydir.h>
 #include <qfile.h>
+#include <qsizepolicy.h>
 #include <QFileInfo>
 #include <QStandardPaths>
 
 #include "lqstring.h"
 
-TMOGUIToneMappingChooser::TMOGUIToneMappingChooser(QWidget *parent) : QScrollArea(parent)
+TMOGUIToneMappingChooser::TMOGUIToneMappingChooser(QWidget *parent, TMOGUIImage *pImg) : QScrollArea(parent)
 {
-    imageWidgetSize = QSize(300, 250);
-    imageFitSize = QSize(290,240);
+    //setAutoFillBackground(true);
+
+    imageWidgetSize = QSize(300, 300);
+    imageFitSize = QSize(290,290);
     listPreview.clear();
     parentWidget = (TMOGUIToneMapping*) parent;
     this->setWidgetResizable(true);
     this->setAlignment(Qt::AlignTop);
-    big_box = new QWidget(this->viewport()); //Q3VBox
+    //this->setBackgroundRole(QPalette::Light);
+    big_box = new QWidget(this->viewport());
     bigBoxLayout = new QVBoxLayout(big_box);
     bigBoxLayout->setAlignment(Qt::AlignTop);
 
@@ -35,29 +39,40 @@ TMOGUIToneMappingChooser::TMOGUIToneMappingChooser(QWidget *parent) : QScrollAre
     //bigBoxLayout->addStretch();
     this->setWidget(big_box);
 
+    if(!pImg){
+        QString previewFileName = "preview-src.tif";
+        imgName = previewFileName;
+        path = new QString(QStandardPaths::writableLocation(QStandardPaths::AppDataLocation));
+        //QString absolutePath = QStandardPaths::locate(QStandardPaths::AppDataLocation, previewFileName);
+        QDir dir(*path);
+        if (!dir.exists())
+            dir.mkpath(*path);
 
-    QString previewFileName = "preview-src.tif";
-    path = new QString(QStandardPaths::writableLocation(QStandardPaths::AppDataLocation));
-    //QString absolutePath = QStandardPaths::locate(QStandardPaths::AppDataLocation, previewFileName);
-    QDir dir(*path);
-    if (!dir.exists())
-        dir.mkpath(*path);
-
-    QString absolutePath = dir.absoluteFilePath(previewFileName);
-    if (QFile::exists(absolutePath))
-        QFile::remove(absolutePath);
+        QString absolutePath = dir.absoluteFilePath(previewFileName);
+        if (QFile::exists(absolutePath))
+            QFile::remove(absolutePath);
 
 
-    QFile::copy(":/resources/images/preview-source.tif", absolutePath);
-    pPreviewImage = new TMOGUIImage(nullptr, this, "Preview-src");
-    pPreviewImage->Open(absolutePath.toStdString().c_str());
-    pPreviewImage->hide();
+        QFile::copy(":/resources/images/preview-source.tif", absolutePath);
+        pPreviewImage = new TMOGUIImage(nullptr, this, "Preview-src");
+        pPreviewImage->Open(absolutePath.toStdString().c_str());
+        pPreviewImage->hide();
+    } else {
+        imgName = pImg->GetImage()->GetFilename();
+        pPreviewImage = new TMOGUIImage(nullptr, this, pImg->GetName(imgName).toStdString().c_str());
+        pPreviewImage->NewSmall(pImg); //TODO check
+        pPreviewImage->setHidden(true);
+
+    }
+    pDefaultImage = pPreviewImage;
+    //displayAll();
 }
 
 TMOGUIToneMappingChooser::~TMOGUIToneMappingChooser()
 {
-    if(pPreviewImage) delete pPreviewImage;
-    if(!listPreview.empty()) listPreview.clear();
+    //if(!listPreview.empty()) listPreview.clear();
+    if(!widgetList.empty()) widgetList.clear();
+    //if(pPreviewImage) delete pPreviewImage;
 }
 
 void TMOGUIToneMappingChooser::AddTMOPreview(TMO* pTMO, int indexLib, int indexTMO)
@@ -65,8 +80,6 @@ void TMOGUIToneMappingChooser::AddTMOPreview(TMO* pTMO, int indexLib, int indexT
 
     lqstring s;
     QString name;
-
-    TMOGUIImage* pImage = nullptr;
 
     try
     {
@@ -77,74 +90,39 @@ void TMOGUIToneMappingChooser::AddTMOPreview(TMO* pTMO, int indexLib, int indexT
         // bad plugin format // plugin version mismatch
     }
 
-    QWidget *widget = new QWidget(big_box);
-    widget->setFixedSize(imageWidgetSize);
-    QVBoxLayout *vLayout = new QVBoxLayout(widget);
-    //QLabel *lbl = new QLabel(s, widget);
-    QPushButton* btn = new QPushButton(s, widget);
+    QPushButton* btn = new QPushButton(name, this);
+    TMOGUIPreviewImage *widget = new TMOGUIPreviewImage(big_box, name, btn, pPreviewImage, pTMO);
+    if(widget->getImageViewSize().height() > 100 && widget->getImageViewSize().width() > 100){
+        widget->getImage()->fitToScreen(widget->getImageViewSize());
+    } else {
+        widget->getImage()->fitToScreen(QSize(300,300));
+    }
 
-    pImage = new TMOGUIImage(nullptr, widget, name.toStdString().c_str());
-    if(pImage->New(pPreviewImage)) return;
-    //pImage->setFixedHeight(pPreviewImage->GetImage()->GetHeight());
-    connect(pImage, SIGNAL(finishTransform()), this, SLOT(finishTransform()));
+    // TODO send widget
+
+    connect(widget, SIGNAL(finishTransform()), this, SLOT(finishTransform()));
     connect(btn, &QPushButton::released, [=]{
         parentWidget->chooseTechnique(indexLib, indexTMO);
     });
 
-    //vLayout->addWidget(lbl);
-    vLayout->addWidget(pImage);
-    vLayout->addWidget(btn);
-
-    pImage->hideAll(true);
-    pImage->fitToWidth(imageFitSize);
-
-    // FIXME Son14 not working
-    if(name != "Son14") listPreview.insert(pImage, pTMO);
+    widgetList.append(widget);
     bigBoxLayout->addWidget(widget, Qt::AlignTop);
+
+
 }
 
-void TMOGUIToneMappingChooser::display()
+void TMOGUIToneMappingChooser::displayAll()
 {
-    TMOImage *pSrc = nullptr;
-    TMOGUITransformation *pTransform = nullptr;
-    QMap<TMOGUIImage*, TMO*>::Iterator i;
-    TMOGUIImage *pImage = nullptr;
-    TMO *pTMO = nullptr;
-
-    listPreviewIt = listPreview.begin();
-    if(listPreviewIt == listPreview.end()) return;
-
-    pImage = listPreviewIt.key();
-    pTMO = listPreviewIt.value();
-
-
-    if(!openPreview(pImage)){
-        pSrc = pImage->GetImage();
-        //pImage->setFixedSize(pSrc->GetWidth(), pSrc->GetHeight());
-        pImage->fitToWidth(imageFitSize);
-        pImage->hideAll(true);
-        delete listPreviewIt.value();
-        listPreview.remove(pImage);
-        display();
-        return;
+    foreach(TMOGUIPreviewImage* w, widgetList){
+        w->display();
     }
-
-    pTransform = pImage->Transform();
-    if (pTransform)
-    {
-        pSrc = pImage->GetImage();
-        pTMO->SetImage(*pSrc);
-
-        //pImage->setFixedSize(pSrc->GetWidth(), pSrc->GetHeight());
-        pImage->fitToWidth(imageFitSize);
-        pTransform->SetTMO(pTMO);
-    }
-
 }
+
 void TMOGUIToneMappingChooser::viewportResizeEvent ( QResizeEvent * e )
 {
     QSize s = e->size();
     backWidth = s.width() - 63;
+
     /*for(TMOGUIImage* pImage : listPreview){
         pImage->fitToWidth(s);
     }*/
@@ -152,11 +130,25 @@ void TMOGUIToneMappingChooser::viewportResizeEvent ( QResizeEvent * e )
 
 void TMOGUIToneMappingChooser::finishTransform()
 {
-    savePreview(listPreviewIt.key());
-    if(!listPreview.empty())
-    {
-        display();
+    //SavePreview(listPreviewIt.key());
+}
+
+void TMOGUIToneMappingChooser::windowChanged(TMOGUIImage* pImage){
+
+    if(pImage){
+
+        if(!pImage->bPreview){
+            pPreviewImage->NewSmall(pImage);
+        } else {
+            pPreviewImage = pDefaultImage;
+        }
+
     }
+    foreach(TMOGUIPreviewImage* w, widgetList){
+        w->setImage(pPreviewImage);
+    }
+
+    displayAll();
 }
 
 int TMOGUIToneMappingChooser::savePreview(TMOGUIImage* pImage)
